requirements.txt
ebooklib>=0.20
beautifulsoup4>=4.12
lxml>=4.9
fastapi>=0.115
uvicorn>=0.30
sqlalchemy>=2.0
python-multipart>=0.0.9
requests
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
bcrypt>=4.0
